move ui config overrides to custom theme
