+++ src/recipe/Baked.Recipe.Service.Application/Theme/Admin/AdminThemeExtensions.cs
@@ -7,5 +7,5 @@ public static class AdminThemeExtensions
 {
     public static AdminThemeFeature Admin(this ThemeConfigurator _,
         IEnumerable<string>? componentExports = default
-    ) => new([.. componentExports ?? []]);
+    ) => new(componentExports ?? []);
 }
