+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/Container.cs
@@ -1,0 +1,8 @@
+using Baked.Ui;
+
+namespace Baked.Test.Theme.Custom;
+
+public record Container : IComponentSchema
+{
+    public List<IComponentDescriptor> Contents { get; init; } = [];
+}
