+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/CustomThemeExtensions.cs
@@ -1,11 +1,10 @@
+using Baked.Test.Theme.Custom;
 using Baked.Theme;
-using Baked.Theme.Admin;
 
 namespace Baked;
 
 public static class CustomThemeExtensions
 {
-    public static CustomThemeFeature Custom(this ThemeConfigurator _,
-        IEnumerable<string>? componentExports = default
-    ) => new([.. componentExports ?? []]);
+    public static CustomThemeFeature Custom(this ThemeConfigurator _) =>
+        new();
 }
