+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/Expected.cs
@@ -1,0 +1,10 @@
+using Baked.Ui;
+
+namespace Baked.Test.Theme.Custom;
+
+public record Expected(string TestId)
+    : IComponentSchema
+{
+    public string TestId { get; set; } = TestId;
+    public bool? ShowDataParams { get; set; }
+}
