+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/Input.cs
@@ -1,0 +1,10 @@
+using Baked.Ui;
+
+namespace Baked.Test.Theme.Custom;
+
+public record Input(string TestId)
+    : IComponentSchema
+{
+    public string TestId { get; set; } = TestId;
+    public string? DefaultValue { get; set; }
+}
