+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/LoginPage.cs
@@ -1,6 +1,6 @@
 ﻿using Baked.Ui;
 
-namespace Baked.Test.ConfigurationOverrider;
+namespace Baked.Test.Theme.Custom;
 
 public record LoginPage(string Path)
     : PageSchemaBase(Path);
