+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/Spec.cs
@@ -1,0 +1,6 @@
+namespace Baked.Test.Theme.Custom;
+
+public record Spec(string Name, IEnumerable<Spec.Link> Links)
+{
+    public record Link(string Title, string Description);
+}
