+++ test/recipe/admin/.baked/components.js
@@ -1,18 +1,17 @@
 import {
-    LazyContainer,
-	LazyExpected,
-	LazyInput,
-	LazyLoginPage,
-	LazyRoutedPage,
-	LazyCardLink,
+    LazyCardLink,
+	LazyContainer,
 	LazyDataPanel,
 	LazyDataTable,
 	LazyDefaultLayout,
 	LazyErrorPage,
+	LazyExpected,
 	LazyFilter,
 	LazyHeader,
-	LazyLanguageSwitcher,
 	LazyIcon,
+	LazyInput,
+	LazyLanguageSwitcher,
+	LazyLoginPage,
 	LazyMenuPage,
 	LazyMessage,
 	LazyModalLayout,
@@ -23,6 +22,7 @@ import {
 	LazyPageTitle,
 	LazyRate,
 	LazyReportPage,
+	LazyRoutedPage,
 	LazySelect,
 	LazySelectButton,
 	LazySideMenu,
@@ -30,20 +30,19 @@ import {
 } from "#components";
 
 export {
-    LazyContainer,
-	LazyExpected,
-	LazyInput,
-	LazyLoginPage,
-	LazyRoutedPage,
-	LazyCardLink,
+    LazyCardLink,
+	LazyContainer,
 	LazyDataPanel,
 	LazyDataTable,
 	LazyDefaultLayout,
 	LazyErrorPage,
+	LazyExpected,
 	LazyFilter,
 	LazyHeader,
-	LazyLanguageSwitcher,
 	LazyIcon,
+	LazyInput,
+	LazyLanguageSwitcher,
+	LazyLoginPage,
 	LazyMenuPage,
 	LazyMessage,
 	LazyModalLayout,
@@ -54,6 +53,7 @@ export {
 	LazyPageTitle,
 	LazyRate,
 	LazyReportPage,
+	LazyRoutedPage,
 	LazySelect,
 	LazySelectButton,
 	LazySideMenu,
